+++ src/features/card/components/CardInput/CardInputOwner/index.tsx
@@ -1,5 +1,5 @@
 import { ChangeEvent } from 'react';
-import { CardInputInterface } from '@/features/card/types/cardInputTypes';
+import { CardInputInterface } from '@/features/card/types/cardTypes';
 import { MAX_LENGTH_CARD_OWNER_NAME } from '@/features/card/constants/maxLength';
 import { Input } from '@/components/atoms/Input';
 import { Label } from '@/components/atoms/Label';
