+++ src/features/card/hooks/useChangeCardNumber.ts
@@ -1,5 +1,5 @@
 import { ChangeEvent, useCallback } from 'react';
-import { CardInputInterface } from '@/features/card/types/cardInputTypes';
+import { CardInputInterface } from '@/features/card/types/cardTypes';
 import { NUMBER_PARSABLE_STRING_PATTERN } from '@/features/card/constants/regExp';
 import { MAX_LENGTH_PIECE_CARD_NUMBER } from '@/features/card/constants/maxLength';
 import { checkIsValidPatternAndLength } from '@/features/card/utils/validator';
