+++ src/features/card/hooks/useChangeOwner.ts
@@ -1,5 +1,5 @@
 import { ChangeEvent, useCallback } from 'react';
-import { CardInputInterface } from '@/features/card/types/cardInputTypes';
+import { CardInputInterface } from '@/features/card/types/cardTypes';
 import { MAX_LENGTH_CARD_OWNER_NAME } from '@/features/card/constants/maxLength';
 import { checkIsValidLength } from '@/features/card/utils/validator';
