+++ src/features/card/hooks/useIsValidCardForm.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { useCard } from '@/features/card/providers/CardProvider';
-import { CardInputInterface } from '@/features/card/types/cardInputTypes';
+import { CardInputInterface } from '@/features/card/types/cardTypes';
 import {
   checkIsValidCardNumber,
   checkIsValidCompanyName,
